requirements.txt
pytest~=8.4.1
requests~=2.32.4
pytest-html
python-dotenv~=1.1.1
jsonschema~=4.25.0